gh-pages: Add .nojekyll file
As of December 2016, GitHub Pages are powered by Jekyll 3.0
which ignores 'node_modules' and 'vendor' by default, regardless
of .gitignore.

While it'll still be added to the git branch, it is not published
over the HTTP service.

Since we use node_modules/qunitjs in sync-gh-pages, we need to
disable this default filter. Do so by creating a '.nojekyll' file
which restores the previous behaviour.